types: re-enable mypy attr-defined, no-any-return (plan 005 stage 3)
Re-enable attr-defined (29) and no-any-return (21).

attr-defined:
- pyproject.toml: add an implicit_reexport=true override for
  tradingview_screener (ships py.typed but doesn't re-export Query/col)
  — resolves 19 import errors.
- unusual_volume/__init__: import Direction/Strength from .classify (source)
  so the re-export is explicit.
- backtester/engine: import _resolve_*_fill from fills.py (source).
- seasonality/rs_breakout/backtester.data: cast .index to DatetimeIndex
  before month/to_period/dayofweek/tz_localize.
- cli: type: ignore[attr-defined] on the _usage_tracked marker attribute.

no-any-return: cast/float()/str() to pin Any returns from untyped libs
(plotly to_html, libsql, jugaad NSE session, json.loads) and from
@indicator/@criterion Callable[..., Any] decorators and pandas
DataFrame|Series unions back to the declared return type. No behavior
changes.

+++ screener/backtester/metrics.py
@@ -38,7 +38,7 @@ def _cagr(equity: pd.Series) -> float:
     # horizon is (N-1)/252 years. Using len(equity)/252 overstated the horizon
     # by one bar and understated CAGR; this matches empyrical's convention.
     years = max((len(equity) - 1) / TRADING_DAYS_PER_YEAR, 1e-9)
-    return (end / start) ** (1.0 / years) - 1.0
+    return float((end / start) ** (1.0 / years) - 1.0)
 
 
 def _max_drawdown(equity: pd.Series) -> float:
